HCM/requirements.txt
allure-python-commons==2.8.40
allure-robotframework==2.8.40
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
chromedriver-py==136.0.7103.113
clipboard==0.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
Faker==8.1.4
idna==2.10
pluggy==0.13.1
pyperclip==1.8.2
python-dateutil==2.8.1
requests==2.25.1
robotframework==4.0.2
robotframework-faker==5.0.0
robotframework-pythonlibcore==2.2.1
robotframework-selenium2library==3.0.0
robotframework-selenium2screenshots==0.8.1
selenium==3.141.0
setuptools==56.2.0
six==1.16.0
text-unidecode==1.3
urllib3==1.26.4
webdriver-manager==3.4.1
wrapt==1.12.1
pandas==1.3.4
xlrd==2.0.1
openpyxl==3.0.10
convertdate==2.4.0
robotframework-pabot==4.0.6
robotframework-datadriver==1.11.2
mailosaur==7.18.0